% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hmp.R
\name{import_hmp}
\alias{import_hmp}
\title{Imports Hapmap genotypic data file}
\usage{
import_hmp(file_path, header = TRUE, sep = "\\t", stringsAsFactors = FALSE)
}
\arguments{
\item{file_path}{Provide the actual path of Hapmap genotypic data file}

\item{header}{by default it will be True}

\item{sep}{by default it will be tab separated}

\item{stringsAsFactors}{by default it will be False}
}
\value{
Hampmap genotypic data
}
\description{
Imports Hapmap genotypic data file
}
\examples{
\donttest{
demo_SNP <- system.file("extdata", "demo_SNP.hmp.txt", package = "geneNR")
hapmap_data <- import_hmp(demo_SNP)
head(hapmap_data)
}
}
