% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vcf.R
\name{import_vcf}
\alias{import_vcf}
\title{Imports VCF (Variant Call Format) data file}
\usage{
import_vcf(file_path)
}
\arguments{
\item{file_path}{Provide the actual path of the VCF file}
}
\value{
A \code{vcfR} object containing the imported data
}
\description{
Imports VCF (Variant Call Format) data file
}
\examples{
\donttest{
demo_SNP <- system.file("extdata", "demo_SNP.vcf", package = "geneNR")
vcf_data <- import_vcf(demo_SNP)
vcf_data
}
}
