% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_rice.R
\docType{data}
\name{sample_data_rice}
\alias{sample_data_rice}
\title{Sample Data}
\format{
A data frame with columns:
\describe{
\item{SNP}{SNP identifier, character.}
\item{Chr}{Chromosome location, character.}
\item{Pos}{Position on the chromosome, numeric.}
\item{traits}{Associated traits, character.}
}
}
\source{
{Basha FTM, Sar P, Bhowmick PK, Mahato A, Bisht DS, Iquebal MA, Chakraborty K, Banerjee A, Verma BC, Bhaduri D, Kumar J, Ngangkham U, Saha S, Priyamedha, Mandal NP, Roy S. Genome-wide association study identified QTLs and genes underlying early seedling vigour in aus rice (Oryza sativa L.). Mol Genet Genomics. 2024 Dec 3;299(1):112. doi: 10.1007/s00438-024-02204-8. PMID: 39625651.}
}
\usage{
sample_data_rice
}
\description{
A dataset containing sample data related to genetic markers and associated traits.
}
\examples{
data(sample_data_rice)    #lazy loading
}
\keyword{datasets}
