% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_wheat.R
\docType{data}
\name{sample_data_wheat}
\alias{sample_data_wheat}
\title{Sample Data}
\format{
A data frame with columns:
\describe{
\item{SNP}{SNP identifier, character.}
\item{Chr}{Chromosome location, character.}
\item{Pos}{Position on the chromosome, numeric.}
\item{traits}{Associated traits, character.}
}
}
\source{
{Generated for demonstration purposes}
}
\usage{
sample_data_wheat
}
\description{
A dataset containing sample data related to genetic markers and associated traits.
}
\examples{
data(sample_data_wheat)    #lazy loading
}
\keyword{datasets}
