% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_wheat_custom.R
\docType{data}
\name{sample_data_wheat_custom}
\alias{sample_data_wheat_custom}
\title{Sample Data}
\format{
A data frame with columns:
\describe{
\item{traits}{Associated traits, character.}
\item{SNP}{SNP identifier, character.}
\item{Chr}{Chromosome location, character.}
\item{start}{Position on the chromosome where search window starts, numeric.}
\item{stop}{Position on the chromosome where search window stops, numeric.}

}
}
\source{
{Generated for demonstration purposes}
}
\usage{
sample_data_wheat_custom
}
\description{
A dataset containing sample data related to genetic markers and associated traits.
}
\examples{
data(sample_data_wheat_custom)    #lazy loading
}
\keyword{datasets}
