% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_wheat_qtl.R
\docType{data}
\name{sample_data_wheat_qtl}
\alias{sample_data_wheat_qtl}
\title{Sample Data}
\format{
A data frame with columns:
\describe{
\item{traits}{Associated traits, character.}
\item{Chr}{Chromosome location, character.}
\item{start}{Position on the chromosome where QTL starts, numeric.}
\item{stop}{Position on the chromosome where QTL stops, numeric.}

}
}
\source{
{Generated for demonstration purposes}
}
\usage{
sample_data_wheat_qtl
}
\description{
A dataset containing sample data related to genetic markers and associated traits.
}
\examples{
data(sample_data_wheat_qtl)    #lazy loading
}
\keyword{datasets}
